motion-game/requirements.txt
PyQt5>=5.15
opencv-python>=4.8
mediapipe>=0.10
pyaudio>=0.2.13
numpy>=1.24
pyobjc-framework-AVFoundation>=10.0; sys_platform == "darwin"
